Делает модульность
