+++ js/const.js
@@ -1,0 +1,27 @@
+import {createRandomIdFromRangeGenerator} from './create-random-id-from-range-generator.js';
+
+const DESCRIPTION = `Описание фотографии.`;
+const COMMENTS = [
+  `Всё отлично!`,
+  `В целом всё неплохо. Но не всё.`,
+  `Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.`,
+  `Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.`,
+  `Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.`,
+  `Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!`
+];
+
+const NAMES = [
+  `Анна`,
+  `Борис`,
+  `Василий`,
+  `Григорий`,
+  `Дмитрий`,
+  `Елена`,
+  `Женя`,
+  `Зинаида`,
+  `Иван`
+];
+const MIN_ID_PHOTO = 1;
+const MAX_ID_PHOTO = 25;
+
+export {DESCRIPTION, COMMENTS, NAMES, MIN_ID_PHOTO, MAX_ID_PHOTO};
